+++ packages/data-designer-slurm/src/data_designer/slurm/serving/resolver.py
@@ -5,6 +5,10 @@
 
 from __future__ import annotations
 
+from dataclasses import dataclass
+
+from pydantic import ValidationError
+
 from data_designer.slurm.config.images import ServingImageInspection
 from data_designer.slurm.config.utils import convert_duration_to_seconds
 from data_designer.slurm.config.vllm import VllmServerConfig
@@ -27,9 +31,16 @@ class VllmServerResolutionError(ValueError):
     """Raised when planner inputs cannot produce a supported vLLM server specification."""
 
 
+@dataclass(frozen=True)
+class VllmServerResolutionContext:
+    """Planner-owned placement and logical endpoint inputs for one vLLM deployment."""
+
+    deployment: ResolvedDeployment
+    logical_endpoint_port: PortClaim
+
+
 def resolve_vllm_server(
-    resolved_deployment: ResolvedDeployment,
-    logical_endpoint_port: PortClaim,
+    context: VllmServerResolutionContext,
 ) -> ResolvedVllmServerDeployment:
     """Resolve one placed vLLM deployment before the complete run plan is assembled.
 
@@ -37,9 +48,15 @@ def resolve_vllm_server(
     baseline behavior required of every admitted vLLM image and therefore does not
     maintain a package-version compatibility matrix.
     """
+    resolved_deployment = context.deployment
+    logical_endpoint_port = context.logical_endpoint_port
     expected_endpoint_id = f"{resolved_deployment.deployment_id}-logical-endpoint"
     if logical_endpoint_port.name != expected_endpoint_id or logical_endpoint_port.role != "logical_endpoint":
         raise VllmServerResolutionError("logical endpoint port must match the resolved deployment")
+    if (logical_endpoint_port.node_index, logical_endpoint_port.port) in {
+        (port.node_index, port.port) for port in resolved_deployment.ports
+    }:
+        raise VllmServerResolutionError("logical endpoint port must not collide with a deployment port")
     authored_deployment = resolved_deployment.authored
     server = authored_deployment.server
     inspection = resolved_deployment.image.inspection_facts
@@ -119,40 +136,43 @@ def resolve_vllm_server(
                     )
                 )
 
-    return ResolvedVllmServerDeployment(
-        deployment_id=resolved_deployment.deployment_id,
-        server_type="vllm",
-        model_alias=authored_deployment.model_alias,
-        model=authored_deployment.model,
-        served_model_name=resolved_deployment.served_model_name,
-        image=resolved_deployment.image,
-        executable_path=inspection.executable_path,
-        node_indices=resolved_deployment.node_indices,
-        gpus_per_node=resolved_deployment.gpus_per_node,
-        topology=resolved_deployment.topology,
-        launch_policy=ResolvedVllmLaunchPolicy(
-            startup_timeout_seconds=convert_duration_to_seconds(server.startup_timeout),
-            distributed_init_timeout_seconds=convert_duration_to_seconds(server.distributed_init_timeout),
-            lead_boot_standoff_seconds=convert_duration_to_seconds(server.lead_boot_standoff),
-            rank_launch_stagger_seconds=convert_duration_to_seconds(server.rank_launch_stagger),
-            readiness_path=server.readiness_path,
-            enable_expert_parallel=server.enable_expert_parallel,
-            queue_backpressure=server.queue_backpressure,
-            extra_args=tuple(server.extra_args),
-            environment=dict(server.environment),
-        ),
-        processes=tuple(processes),
-        readiness_probes=tuple(probes),
-        backend_endpoints=tuple(backends),
-        logical_endpoint=ResolvedLogicalEndpoint(
-            endpoint_id=logical_endpoint_port.name,
+    try:
+        return ResolvedVllmServerDeployment(
+            deployment_id=resolved_deployment.deployment_id,
+            server_type="vllm",
             model_alias=authored_deployment.model_alias,
+            model=authored_deployment.model,
             served_model_name=resolved_deployment.served_model_name,
-            node_index=logical_endpoint_port.node_index,
-            port=logical_endpoint_port.port,
-            backend_ids=tuple(backend.backend_id for backend in backends),
-        ),
-    )
+            image=resolved_deployment.image,
+            executable_path=inspection.executable_path,
+            node_indices=resolved_deployment.node_indices,
+            gpus_per_node=resolved_deployment.gpus_per_node,
+            topology=resolved_deployment.topology,
+            launch_policy=ResolvedVllmLaunchPolicy(
+                startup_timeout_seconds=convert_duration_to_seconds(server.startup_timeout),
+                distributed_init_timeout_seconds=convert_duration_to_seconds(server.distributed_init_timeout),
+                lead_boot_standoff_seconds=convert_duration_to_seconds(server.lead_boot_standoff),
+                rank_launch_stagger_seconds=convert_duration_to_seconds(server.rank_launch_stagger),
+                readiness_path=server.readiness_path,
+                enable_expert_parallel=server.enable_expert_parallel,
+                queue_backpressure=server.queue_backpressure,
+                extra_args=tuple(server.extra_args),
+                environment=dict(server.environment),
+            ),
+            processes=tuple(processes),
+            readiness_probes=tuple(probes),
+            backend_endpoints=tuple(backends),
+            logical_endpoint=ResolvedLogicalEndpoint(
+                endpoint_id=logical_endpoint_port.name,
+                model_alias=authored_deployment.model_alias,
+                served_model_name=resolved_deployment.served_model_name,
+                node_index=logical_endpoint_port.node_index,
+                port=logical_endpoint_port.port,
+                backend_ids=tuple(backend.backend_id for backend in backends),
+            ),
+        )
+    except ValidationError as error:
+        raise VllmServerResolutionError("planner inputs produced an inconsistent vLLM server specification") from error
 
 
 def _resolve_rendezvous(
